backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
argon2-cffi>=23.1.0
pytest>=7.4.0
scipy>=1.7.0
beautifulsoup4>=4.9.0
lxml>=4.6.0